services/pricing-booking-py/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
httpx==0.28.1
SQLAlchemy==2.0.32
python-ulid==2.7.0
PyJWT==2.9.0


